% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMpack-Analyses.r
\name{convergeIPM}
\alias{convergeIPM}
\title{Iteratively find best number of bins for a given tolerance and metric}
\usage{
convergeIPM(growObj, survObj, fecObj, nBigMatrix, minSize, maxSize,
  discreteTrans = 1, integrateType = "midpoint", correction = "none",
  preCensus = TRUE, tol = 1e-04, binIncrease = 5, chosenBin = 1,
  response = "lambda")
}
\arguments{
\item{growObj}{An object of class \code{growObj}}

\item{survObj}{An object of class \code{survObj}}

\item{fecObj}{An object of class \code{fecObj}}

\item{nBigMatrix}{A numeric specifying the initial number of bins
used in the matrix. This number will be \emph{increased} in the function}

\item{minSize}{A numeric specifying the lower size bound for the 
big matrix's meshpoints}

\item{maxSize}{A numeric specifying the upper size bound for the 
big matrix's meshpoints}

\item{discreteTrans}{A matrix of discrete transitions. If there are none,
specify as 1}

\item{integrateType}{The type of function to integrate over. The default
is \code{midpoint}. This uses the probability density function.
\code{cumul} is the other option. This uses the cumulative density function.}

\item{correction}{The type of correction to use for eviction. 
The first option is \code{constant} which will multiply every column 
of the IPM by a constant sufficient to adjust values to those
predicted for survival at that size. The second option is
\code{discretizeExtremes} which will place all transitions to sizes
smaller than \code{minSize} into the smallest bin, and transitions to
sizes larger than \code{maxSize} into the largest bin.}

\item{preCensus}{A logical defining whether the \code{fecObj} is 
pre or post-census. The default is pre-census.}

\item{tol}{The desired tolerance level for \code{response}.}

\item{binIncrease}{The number of bins to add in each successive 
iteration.}

\item{chosenBin}{The desired bin for which life expectancy should
be assessed. The default is the first.}

\item{response}{The variable to test for convergence. The options
are \code{lambda}, \code{R0}, and \code{lifeExpect}. The latter will
calculate the life expectancy for individuals in \code{chosenBin}.}
}
\value{
A list with the following elements: 
\itemize{
  \item{\code{binIncrease}{ - The number of bins used to increase the 
                           matrix size}}
  \item{\code{Pmatrix}{ - The final \code{Pmatrix} if only life expectancy
                       is being considered}}
  \item{\code{IPM}}{ - The final IPM reached at convergence}
  \item{\code{R0}}{ - The final \code{R0}}
  \item{\code{lambda}}{ - The final \code{lambda}}
  \item{\code{LE}}{ - The final vector of life expectancies}
}
}
\description{
Iteratively increases the number of bins specified in
\code{binIncrease} until the difference in the chosen measure falls
below the selected tolerance level.
}
\details{
Different choices for responses will yield different values.
The pattern of change in lambda (or other response variables) can
be complex, so it is advisable to start with large binIncrease and small
tolerance, and then once one knows a general idea of how big the matrix 
needs to be, run the function again with a smaller binIncrease but start 
it closer to the goal.

For the life expectancy option, if discrete stages are included via 
discreteTrans then if chosenBin=1, this function will use the first 
discrete bin.
}
\note{
This code was modified from original code by Melissa Eitzel.
}
\examples{
dff<-generateData()
gr1<-makeGrowthObj(dff)
sv1<-makeSurvObj(dff)
fv1<-makeFecObj(dff,Transform="log")
res <- convergeIPM(growObj=gr1, 
                   survObj=sv1, fecObj=fv1, 
                   nBigMatrix=10, minSize=-2,
                   maxSize=15,discreteTrans = 1, 
                   integrateType = "midpoint", 
                   correction = "none", 
                   preCensus = TRUE, tol=1e-3,binIncrease=10)
res <- convergeIPM(growObj=gr1, 
                   survObj=sv1, fecObj=fv1, 
                   nBigMatrix=10, minSize=-2,
                   maxSize=15,discreteTrans = 1, 
                   integrateType = "midpoint", 
                   correction = "none", 
                   preCensus = TRUE, tol=1e-3,
                   binIncrease=10, response="R0")

res <- convergeIPM(growObj=gr1, survObj=sv1, fecObj=fv1, 
                   nBigMatrix=10, minSize=-2,
                   maxSize=15,discreteTrans = 1, 
                   integrateType = "midpoint", 
                   correction = "none", 
                   preCensus = TRUE, tol=1e-3,binIncrease=10, 
                   response="lifeExpect")

}
\seealso{
\code{\link{diagnosticsPmatrix}}
}
\author{
C. Jessica E. Metcalf, 
Sean M. McMahon, Roberto Salguero-Gomez,
Eelke Jongejans & Cory Merow.
}
