\name{fecObj-class}
\Rdversion{1.1}
\docType{class}
\alias{fecObj-class}

\title{Class "fecObj"}
\description{
A class object description
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("fecObj", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{

   "fecConstants"      
"offspringSplitter"  "meanOffspringSize" "varOffspringSize"  "fecByDiscrete"     "Transform"
    
    \item{\code{fitFec}:}{Object of class \code{"list"} ~~ }
    \item{\code{fecConstants}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{offspringSplitter}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{meanOffspringSize}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{varOffspringSize}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{fecByDiscrete}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{Transform}:}{Object of class \code{"character"} ~~ }
   
  }
}
\section{Methods}{
No methods defined with class "fecObj" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
C. Jessica E. Metcalf, Sean McMahon, Roberto Salguero-Gomez & Eelke Jongejans.
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("fecObj")
}
\keyword{classes}
