\name{IPMpack-package}
\alias{IPMpack-package}
\alias{IPMpack}
\docType{package}
\title{
Construction and analysis of integral projection models and associated measures of population growth, structure, perturbations (sensitivities and elasticities), overall population dynamics, age-specific metrics, etc. 
}
\description{
IPM package, a series of analytical tools for building, diagnosing, and projecting populations models based on continuous and descrete vital rates.}
\details{
\tabular{ll}{
Package: \tab IPMpack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-21-03\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Depends: \tab MCMCglmm, Matrix, MASS, nlme\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
The IPMpack team: C. Jessica E. Metcalf, Sean McMahon, Roberto Salguero-Gomez & Eelke Jongejans.
Maintainer: ipmpack-users@lists.r-forge.r-project.org; ipmpack@gmail.com
User mailing list: https://lists.r-forge.r-project.org/mailman/listinfo/ipmpack-users 
}
\references{
%Insert here the MEE paper when(/if) it comes out ROB TO DO
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
%~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
