\name{SilwoodDataClean.csv}
\alias{SilwoodDataClean.csv}
\docType{data}
\title{
Silwood Monocarpic Plant Demographic Data
}
\description{
Data-frame containing size and survival or flowering from two years of
data from species from Silwood Park. 
}
\usage{data(/Users/cjessicametcalf/Documents/IPM/data/SilwoodDataClean.csv)}
\format{
  The format is:
 chr "/Users/cjessicametcalf/Documents/IPM/data/SilwoodDataClean.csv"
}
\details{
data-frame with headings exactDate (date measurement) exactDateNext
(date next measurement) id (unique plant id) Species (species name) Site
(location within Silwood park where plant was measured) rtcr (root crown
diameter, measured with callipers) rtcrNext (root crown diameter at next
census time) ll (length of longest leaf) llNext (length of longest leaf
at next census) rosetteDiam (rosette Diameter) rosetteDiamNext (rosette
diameter at next census) flowered (0 for not flowered, 1 if did) surv (0
if died, one if survived, NAs if not know). 
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(/Users/cjessicametcalf/Documents/IPM/data/SilwoodDataClean.csv)

}
\keyword{datasets}
